requirements.txt

requests>=2.25.0
Pillow>=8.0.0
tqdm>=4.50.0
cairosvg>=2.5.0
python-pptx>=0.6.0